pyproject.toml
[project]
name = "googleadk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.118.3",
    "google-adk[extensions]>=1.19.0",
    "litellm>=1.80.7",
    "pydantic>=2.12.5",
    "greenlet>=3.0.0",
    "scikit-learn>=1.7.2",
    "python-dotenv>=1.2.1",
    "motor>=3.7.1",
    "pydantic-settings>=2.12.0",
    "langchain-community>=0.4.1",
    "wikipedia>=1.4.0",
    "ruff>=0.14.8",
    "requests>=2.32.5",
    "mem0ai>=1.0.2",
    "reportlab>=4.4.9",
]
